refactor(env): fix the env order to be more logical
The base env should be overriden by any other type of env. There is a lil exception for the
`NODE_ENV` one that is required for the env vars loading.

BREAKING CHANGE: Env vars won't be loaded in the same order. That said, it is probably the order you
would expect.